Use linaro-django-pagination instead of django-pagination.
django-pagination hasn't been update in 4 years, despite many pull requests. It
looks like the author of the library abandoned it without even bothering to
delegate responsibility to someone else, which makes it a bad choice for a
dependency if we want to stay up-to-date. References:

https://github.com/ericflo/django-pagination/pull/21
https://github.com/ericflo/django-pagination/pull/22

linaro-django-pagination (https://github.com/zyga/django-pagination) is better
maintained and should be used instead.