Fix category toggle JS collapse
On my browser, the toggle didn't work well. It wouldn't properly toggle
the +/- icon and it wouldn't remember the toggle state in the cookies. I
found out it was becuase of the `== ''` test: for, me, the class was
`undefined` so the test never passed.

To fix it, I use the more robust `hasClass()` jQuery function.