Allow jQuery URL override
Allow jQuery's URL (currently a hardcoded URL to a cloudflare CDN) to be
overridden by the forum project.

I did this because I was intermittently getting SSL warnings because of
mixed contents. I felt that self-hosting was the easiest way to solve
this problem, so I added this setting.

As far as I can tell, this is the only hardcoded outside reference we
have, so it should be the only case where we need such an override.